+++ src/main/java/com/trend/cloudone/amaas/AMaasErrorCode.java
@@ -8,32 +8,32 @@ public enum AMaasErrorCode {
      * Scan file cannot be found.
      */
     MSG_ID_ERR_FILE_NOT_FOUND("MSG_ID_ERR_FILE_NOT_FOUND", "Failed to open file. No such file or directory %s."),
-    
+
     /**
      * Have no permission to open the scan file.
      */
     MSG_ID_ERR_FILE_NO_PERMISSION("MSG_ID_ERR_FILE_NO_PERMISSION", "Failed to open file. Permission denied to open %s."),
-    
+
     /**
      * Incorrect region is given.
      */
     MSG_ID_ERR_INVALID_REGION("MSG_ID_ERR_INVALID_REGION", "%s is not a supported region, region value should be one of %s"),
-    
+
     /**
     * Cloudone credetial is not given.
     */
-    MSG_ID_ERR_MISSING_AUTH("MSG_ID_ERR_MISSING_AUTH","Must provide an API key to use the client."),
-    
+    MSG_ID_ERR_MISSING_AUTH("MSG_ID_ERR_MISSING_AUTH", "Must provide an API key to use the client."),
+
     /**
      * An GRPC error was reported.
      */
     MSG_ID_GRPC_ERROR("MSG_ID_GRPC_ERROR", "Received gRPC status code: %d, msg: %s."),
-    
+
     /**
      * Authentication failed using the given Cloudone credential.
      */
     MSG_ID_ERR_KEY_AUTH_FAILED("MSG_ID_ERR_KEY_AUTH_FAILED", "Authorization key cannot be authenticated."),
-    
+
     /**
      * Java SDK client received an unexpected interrupt.
      */
@@ -63,7 +63,7 @@ public enum AMaasErrorCode {
      * @param errorCode error code to be used
      * @param message template to be used.
      */
-    AMaasErrorCode(String errorCode, String message) {
+    AMaasErrorCode(final String errorCode, final String message) {
         this.errorCode = errorCode;
         this.message = message;
     }
@@ -89,7 +89,7 @@ public enum AMaasErrorCode {
      * @param params objects to be used to format the template.
      * @return message prepared from the templates using the parameters.
      */
-    public String getMessage(Object... params) {
+    public String getMessage(final Object... params) {
         return String.format(this.message, params);
     }
 }