+++ src/test/java/com/trend/cloudone/amaas/TestRunner.java
@@ -6,15 +6,17 @@ import org.junit.runner.Result;
 import org.junit.runner.notification.Failure;
 
 
-public class TestRunner {
+public final class TestRunner {
+   private TestRunner() {
+   }
 
-   public static void main(String[] args) {
+   public static void main(final String[] args) {
       Result result = JUnitCore.runClasses(TestSuite.class);
-		
+
       for (Failure failure : result.getFailures()) {
          System.out.println(failure.toString());
       }
-		
+
       System.out.println(result.wasSuccessful());
    }
 }