+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # CHANGELOG
 
-## 1.0.0 - 2023-11-7
+## 1.1.0 - 2024-04-03
+* Update protos
+* Enable PML (Predictive Machine Learning) detection and smart feedback
+* Enable bulk mode
+* Enable India region
+* Support for scanning large files (over 2GB)
+
+## 1.0.1 - 2023-11-23
+* Update to latest code
+
+## 1.0.0 - 2023-11-21
 * Initial release